fix(python): Fix dtype parameter in `pandas_to_pyseries` function

+++ py-polars/polars/series/series.py
@@ -346,7 +346,7 @@ class Series:
         elif _check_for_pandas(values) and isinstance(
             values, (pd.Series, pd.Index, pd.DatetimeIndex)
         ):
-            self._s = pandas_to_pyseries(name, values)
+            self._s = pandas_to_pyseries(name, values, dtype=dtype)
 
         elif _is_generator(values):
             self._s = iterable_to_pyseries(name, values, dtype=dtype, strict=strict)
